requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy[asyncio]
asyncpg
pydantic_settings
alembic
psycopg2
black
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart